Add auth to user routes

+++ src/handlers/user_handler.rs
@@ -1,4 +1,8 @@
-use crate::{error::Result, models::User, requests::user::CreateUser};
+use crate::{
+    error::{Result, ServiceError},
+    models::User,
+    requests::user::CreateUser,
+};
 use actix_web::{
     delete, get, post, put,
     web::{Data, Json, Path, ServiceConfig},
@@ -20,19 +24,27 @@ pub async fn register(data: Json<CreateUser>, pool: Data<PgPool>) -> Result<Http
 }
 
 #[get("/user/{id}")]
-pub async fn find(id: Path<i64>, pool: Data<PgPool>) -> Result<HttpResponse> {
+pub async fn find(id: Path<i64>, pool: Data<PgPool>, _: User) -> Result<HttpResponse> {
     let user = User::find(&pool, id.into_inner()).await?;
 
     Ok(HttpResponse::Ok().json(user))
 }
 
 #[put("/user/{id}")]
-pub async fn update(_: Path<i64>, _: Json<CreateUser>) -> impl Responder {
-    HttpResponse::Ok().finish()
+pub async fn update(id: Path<i64>, auth_user: User, _: Json<CreateUser>) -> impl Responder {
+    if auth_user.id != id.clone() {
+        Err(ServiceError::Forbidden)
+    } else {
+        Ok(HttpResponse::Ok().finish())
+    }
 }
 
 #[delete("/user/{id}")]
-pub async fn delete(id: Path<i64>, pool: Data<PgPool>) -> Result<HttpResponse> {
+pub async fn delete(id: Path<i64>, auth_user: User, pool: Data<PgPool>) -> impl Responder {
+    if auth_user.id != id.clone() {
+        return Err(ServiceError::Forbidden);
+    }
+
     User::delete(&pool, id.into_inner()).await?;
 
     Ok(HttpResponse::NoContent().finish())
